Improve output in case of mismatch.  This test fails for me: FAIL!  : KTcpSocketTest::states() Compared values are not the same    Actual ((int)s->state()): 6    Expected ((int)KTcpSocket::UnconnectedState): 0    Loc: [/d/kde/src/4/kdelibs/kdecore/tests/ktcpsockettest.cpp(225)] CCMAIL: ahartmetz@gmail.com
svn path=/trunk/KDE/kdelibs/; revision=752534

+++ kdecore/tests/ktcpsockettest.cpp
@@ -222,7 +222,7 @@ void KTcpSocketTest::states()
     s->close();
     //What happens is that during waitForReadyRead() the write buffer is written out
     //completely so that the socket can shut down without having to wait for writeout.
-    QCOMPARE(s->state(), KTcpSocket::UnconnectedState);
+    QCOMPARE((int)s->state(), (int)KTcpSocket::UnconnectedState);
 
     //Now again an internal connection
     invokeOnServer("states");
